Add fix-react-globe script
